# Improve cache files performances