--- a/lib/historyDiff.ml
+++ b/lib/historyDiff.ml
@@ -49,7 +49,7 @@ let create_history_dirs conf fname =
     let dirs =
       [ history_d conf; String.make 1 fname.[0]; String.make 1 fname.[1] ]
     in
-    Mutil.mkdir_p (List.fold_left Filename.concat "" dirs)
+    File.create_dir ~parent:true (List.fold_left Filename.concat "" dirs)
 
 (* ************************************************************************ *)
 (*  [Fonc] write_history_file : config -> string -> gen_record -> unit      *)