My PR for the search engine is too large now so I start to split it in atomic ones. 

The main purpose of this PR is to improve the performance of the script `cache_files`. The new script is twice faster. The PR includes also:
1. A tiny library for file manipulations in `lib/util/file.mli`. 
2. A better progress bar. The previous one was buggy, can only be used on `stderr` and cannot be used in hot loops. 
3. A full rewrite of the script `cache_file`. Please, let me know if it works fine for you. 

Some general remarks:
1. The library `File` includes a new function `File.create_dir` which can be used to create new directories without raising an error if the directory already exists. The function can also create parents if they do not exists but it is not as good as `mkdir -p` in your terminal. For instance, `File.create_dir ~parent:true "foo/../bar"` will fail because of the dot-dot in the middle. I think that the correct way to implement it is to normalize the input path. On `Unix`, we can use `realpath` of the libc but the OCaml Unix library do not expose it. I think that this function is also useful for security purposes. For instance, I believe that `Secure.check` is a real sieve because of it is implementation of `Secure.decompose` (I didn't check it, I may be wrong).
2. In plenty of locations of the codebase, the backtrace of OCaml is destroyed by the following anti-pattern:
```ocaml 
(* Get some resource *)
let resource = open () in
try 
  (* do something that can raise exceptions *)
  close (); 
with 
| exn -> 
   close (); 
   raise exn
```
The OCaml runtime cannot understand that the exception have been raised in the try block and you will only got the end of the backtrace (which is, most of the times, the least interesting part of it). It is error-prone to maintain this pattern by hand, and actually I found places in `gwd` where resources are not freed for all the exceptions. If you want to save the backtrace, you have to it by hand: 
```ocaml 
(* Get some resource *)
let resource = open () in
try 
  (* do something that can raise exceptions *)
  close (); 
with 
| exn -> 
   close (); 
   let bt = Printexc.get_raw_backtrace () in
   Printexc.raise_with_backtrace exn bt
```
 Unfortunately, OCaml has no `finally` blocks like Java or Python, so you have to it yourself. Since OCaml 4.08, there is a new function `Fun.protect` which essentially do the right thing for you. The previous code can be rewrite like this:
```ocaml 
(* Get some resource *)
let resource = open () in
Fun.protect ~finally:close @@ fun () -> 
  (* do something that can raise exceptions *)
 ```

The `File` module contains two specialized functions of `Fun.protect` for files and directories. I recommend that you use them.