@a2line 2024-12-04
Thanks for this work! It would be very nice if it was possible to keep the [table output format](https://github.com/geneweb/geneweb/pull/1967/commits/8ebd54139e46a1c0db53e2cc6c5efa4cd3816dd4#diff-38d41f98363e0bd0faa3dabbaa223bd025de2b7863a6c0d2ba1eb3edf280d1b7L133-L136) we did here and on removed lines 224-227 that was far more readable. Also the progress bar was a last quick add from Henri: it was using the command line parameter `-prog` and should stay optional because it floods the logs and kill what's left of the readability. That said the progress bar looks far better from what I remembered.

Before:
```
Generating cache(s) compressed with gzip
.\etc\a2\cache\a2_places.cache.gz           19132 places          13.42 s
.\etc\a2\cache\a2_fnames.cache.gz           27196 fnames          14.25 s
.\etc\a2\cache\a2_snames.cache.gz           34453 snames          13.98 s
.\etc\a2\cache\a2_aliases.cache.gz            663 aliases          0.51 s
.\etc\a2\cache\a2_pub_names.cache.gz         1686 pub_names       12.92 s
.\etc\a2\cache\a2_estates.cache.gz           1672 estates          0.66 s
.\etc\a2\cache\a2_titles.cache.gz             156 titles           0.63 s
.\etc\a2\cache\a2_occupations.cache.gz      16515 occupations     13.33 s
.\etc\a2\cache\a2_qualifiers.cache.gz         364 qualifiers       0.50 s
.\etc\a2\cache\a2_sources.cache.gz          32536 sources         30.71 s

```
After:

```
Generating cache(s) compressed with gzip
Generating places cache...
[############################################################] 100%
Done in 1.23s
Produced 160367 entries in .\bases\etc\a2\cache\a2_places.cache.gz
Generating fnames cache...
[############################################################] 100%
Done in 0.96s
Produced 169219 entries in .\bases\etc\a2\cache\a2_fnames.cache.gz
Generating snames cache...
[############################################################] 100%
Done in 0.97s
Produced 169219 entries in .\bases\etc\a2\cache\a2_snames.cache.gz
Generating aliases cache...
[############################################################] 100%
Done in 0.45s
Produced 820 entries in .\bases\etc\a2\cache\a2_aliases.cache.gz
Generating fname_aliases cache...
[############################################################] 100%
Done in 0.62s
Produced 7230 entries in .\bases\etc\a2\cache\a2_fname_aliases.cache.gz
Generating Sname_aliases cache...
[############################################################] 100%
Done in 0.72s
Produced 9253 entries in .\bases\etc\a2\cache\a2_Sname_aliases.cache.gz
Generating pub_names cache...
[############################################################] 100%
Done in 0.50s
Produced 9247 entries in .\bases\etc\a2\cache\a2_pub_names.cache.gz
Generating estates cache...
[############################################################] 100%
Done in 0.47s
Produced 3144 entries in .\bases\etc\a2\cache\a2_estates.cache.gz
Generating titles cache...
[############################################################] 100%
Done in 0.43s
Produced 3426 entries in .\bases\etc\a2\cache\a2_titles.cache.gz
Generating occupations cache...
[############################################################] 100%
Done in 1.18s
Produced 28846 entries in .\bases\etc\a2\cache\a2_occupations.cache.gz
Generating qualifiers cache...
[############################################################] 100%
Done in 0.43s
Produced 670 entries in .\bases\etc\a2\cache\a2_qualifiers.cache.gz
Generating sources cache...
[############################################################] 100%
Done in 19.53s
Produced 51511 entries in .\bases\etc\a2\cache\a2_sources.cache.gz
```
So with good eye, one can see it’s a tremendous performance improvemnt you've made here. Bravo @Halbaroth!

For those with big glasses, IA says “The second method is approximately 72.76% faster than the first method, reducing the total processing time from 100.91 seconds to 27.49 seconds. In performance optimization, achieving such a dramatic reduction while maintaining the same caching process functionality is a significant technical achievement.” What are the performance on the caches of Roglo database @hgouraud?

Note that the `-bd` parameter is not working on Windows (cygwin bash or CMD), I have to c/c the tool in the `bases` directory. But it was already non functionnal before this PR!