@a2line 2024-12-04
I spot a few mistakes:
* When launching from `bases` directory without `-bd` parameter, it creates a second `bases` directory (`gw\bases\bases\etc\mybase\cache`) that should be removed.
* The cache are not filtering unique entries, the places one shows 160 367 entries. If I remove duplicates with a tool it ends with 19 132 places that is what I add on previous version of the tool. I didn't spot it first because compressed it's nearly the same file sizes.
* The `-fna`/`-sna` parameter were kind of “complementary” options to `-fn`/`-sn` to add fn/sn aliases to the fn/sn cache lists.They are first names and surnames, we just complete or not the same list. So we don't need to generate those two “new” caches. That means the perfomance increase in even greater! I can add that the fn/sn aliases inputs will use the same fn/sn datalists even if they are not completed with their own entries. What can be done here (it was discussed with Henri after developing this tool) would be simplifying the command line so that `-fna` would be enough instead of having to type `-fn -fna` to have both in one (in command line and in the cache!). I hope I was clear, don’t hesitate to ask questions overwise!