@Halbaroth 2024-12-05
I pushed a new version:
1. I restored the output with tabulation.
2. Hide completely the progress bar if `-prog` is absent. 
3. Deduplicate data. Good catch for this one. The OCaml API of hash tables is very weird and I forgot it. `Hashtbl.add` do not replace a binding if the key is already present (like in practically all other implementations) and `Hashtbl.replace` is the function we really want.

For the bug with `-bd`, honestly I don't understand how the previous script was supposed to be called, may be @hgouraud could explain us how he calls it? 
In my opinion, it would be far simpler to have three arguments:
1. `-output` for the output directory 
2. `-bd`  for the root of databases 
3. the main argument is the name of the database itself.