@Halbaroth 2024-12-05
I add a module `compat` to backport modules and functions of the stdlib that are not available on old OCaml version. Actually my functions `with_in_open` and `with_out_open` in `File` have been added in `In_channel` and `Out_channel` of the stdlib, so I copied them in `compat` and use them.