@Halbaroth 2024-12-05
I bench a last time the script. The next bottle neck to remove comes from `Iovalue`. Basically, Iovalue implements marshaling of data structures to store them in the database. The script spends 25% of its run in the function `Iovalue.input_block`. IO accesses are slow, so it makes sense that the script spends a lot of time here, but the implementation of this module is written with unsafe OCaml (there are a lot of Obj.magic to bypass the typechecker, which is necessary for a marshaling module) and actually there is already such a module in the stdlib named `Marshal`. Geneweb should use it because:
1. `Iovalue` relies on the memory representation of OCaml values but the OCaml project does not give any guarantee on the stability of this ABI. In practice, the memory representation does not change often. 
2. Such a low-level function cannot be written efficiently in OCaml. The implementation of the stdlib is written in C. 

The documentation of the module claims that these functions are identical to the functions exposed by the Marshal module, which is incorrect. Changing this module is not easy because the current format does not seem to be compatible with Marshal. Still, it would be valuable to change this, as it should improve all the IO accesses.